refactor: split modules.ts into imageUrl.ts (pure functions) and chromeApi.ts (Chrome API wrappers)
